wip_scripts/media_names_format/pyproject.toml
[tool.poetry]
name = "media_names_format"
version = "0.1.0"
description = "A script that renames media files based on its metadata (EXIF and such)."
authors = ["Michal Bultrowicz <michal.bultrowicz@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
exif = "^1.3.5"
ffmpeg-python = "^0.2.0"
hachoir = "^3.2.0"
pillow = "^9.4.0"

[tool.poetry.dev-dependencies]
pdbpp = "^0.10.3"
pytest = "^7.1.1"
ptpython = "^3.0.20"
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
